Separation of the coeff from QobjEvo
