+++ qutip/qobjevo.py
@@ -57,148 +57,6 @@ safePickle = [False]
 if sys.platform == 'win32':
     safePickle[0] = True
 
-try:
-    import cython
-    use_cython = [True]
-except:
-    use_cython = [False]
-
-
-def proj(x):
-    if np.isfinite(x):
-        return (x)
-    else:
-        return np.inf + 0j * np.imag(x)
-
-
-str_env = {
-    "sin": np.sin,
-    "cos": np.cos,
-    "tan": np.tan,
-    "asin": np.arcsin,
-    "acos": np.arccos,
-    "atan": np.arctan,
-    "pi": np.pi,
-    "sinh": np.sinh,
-    "cosh": np.cosh,
-    "tanh": np.tanh,
-    "asinh": np.arcsinh,
-    "acosh": np.arccosh,
-    "atanh": np.arctanh,
-    "exp": np.exp,
-    "log": np.log,
-    "log10": np.log10,
-    "erf": scipy.special.erf,
-    "zerf": scipy.special.erf,
-    "sqrt": np.sqrt,
-    "real": np.real,
-    "imag": np.imag,
-    "conj": np.conj,
-    "abs": np.abs,
-    "norm": lambda x: np.abs(x)**2,
-    "arg": np.angle,
-    "proj": proj,
-    "np": np,
-    "spe": scipy.special}
-
-
-class _file_list:
-    """
-    Contain temp a list .pyx to clean
-    """
-    def __init__(self):
-        self.files = []
-        self.clean_old()
-
-    def add(self, file_):
-        self.files += [file_ + ".pyx"]
-
-    def clean(self):
-        import os
-        to_del = []
-        for i, file_ in enumerate(self.files):
-            try:
-                os.remove(file_)
-                to_del.append(i)
-            except Exception:
-                if not os.path.isfile(file_):
-                    to_del.append(i)
-
-        for i in to_del[::-1]:
-            del self.files[i]
-
-    def clean_old(self, max_age=qset.max_age):
-        # On importing Qutip, remove temp files left-over for more than 1 week.
-        import glob
-        import time
-        root = qset.tmproot
-        bases = ["td_Qobj_single_str",
-                 "qobjevo_compiled_coeff_",
-                 "cqobjevo_compiled_coeff_"]
-        for base in bases:
-            for file in glob.glob(os.path.join(root, base + "*")):
-                try:
-                    age = (time.time() - os.path.getctime(file)) / 3600
-                    if age > max_age:
-                        os.remove(file)
-                except:
-                    pass
-                    # Warn that the  file could not be deleted?
-
-
-coeff_files = _file_list()
-atexit.register(coeff_files.clean)
-
-
-class _StrWrapper:
-    def __init__(self, code):
-        self.code = "_out = " + code
-
-    def __call__(self, t, args={}):
-        env = {"t": t}
-        env.update(args)
-        exec(self.code, str_env, env)
-        return env["_out"]
-
-class _CubicSplineWrapper:
-    # Using scipy's CubicSpline since Qutip's one
-    # only accept linearly distributed tlist
-    def __init__(self, tlist, coeff, args=None):
-        self.coeff = coeff
-        self.tlist = tlist
-        try:
-            use_step_func = args["_step_func_coeff"]
-        except KeyError:
-            use_step_func = 0
-        if use_step_func:
-            self.func = interp1d(
-                self.tlist, self.coeff, kind="previous",
-                bounds_error=False, fill_value=0.)
-        else:
-            self.func = CubicSpline(self.tlist, self.coeff)
-
-    def __call__(self, t, args={}):
-        return self.func([t])[0]
-
-class _StateAsArgs:
-    # old with state (f(t, psi, args)) to new (args["state"] = psi)
-    def __init__(self, coeff_func):
-        self.coeff_func = coeff_func
-
-    def __call__(self, t, args={}):
-        return self.coeff_func(t, args["_state_vec"], args)
-
-
-# %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-class StateArgs:
-    """Object to indicate to use the state in args outside solver.
-    args[key] = StateArgs(type, op)
-    """
-    def __init__(self, type="Qobj", op=None):
-        self.dyn_args = (type, op)
-
-    def __call__(self):
-        return self.dyn_args
 
 # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 # object for each time dependent element of the QobjEvo
@@ -206,6 +64,7 @@ class StateArgs:
 # get_coeff : a callable that take (t, args) and return the coeff at that t
 # coeff : The coeff as a string, array or function as provided by the user.
 # type : flag for the type of coeff
+"""
 class EvoElement:
     def __init__(self, qobj, get_coeff, coeff, type):
         self.qobj = qobj
@@ -226,7 +85,21 @@ class EvoElement:
             return self.coeff
         if i == 3:
             return self.type
+"""
 
+EvoElement = namedtuple("EvoElement", "qobj coeff")
+
+# %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+class StateArgs:
+    """Object to indicate to use the state in args outside solver.
+    args[key] = StateArgs(type, op)
+    """
+    def __init__(self, type="Qobj", op=None):
+        self.dyn_args = (type, op)
+
+    def __call__(self):
+        return self.dyn_args
+# %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 class QobjEvo:
     """A class for representing time-dependent quantum objects,
@@ -490,81 +363,40 @@ class QobjEvo:
         if isinstance(Q_object, list) and len(Q_object) == 2:
             if isinstance(Q_object[0], Qobj) and not isinstance(Q_object[1],
                                                                 (Qobj, list)):
-                # The format is [Qobj, f/str]
+                # The format is [Qobj, coeff]
                 Q_object = [Q_object]
 
         op_type = self._td_format_check_single(Q_object, tlist)
         self.ops = []
 
-        if isinstance(op_type, int):
-            if op_type == 0:
-                self.cte = Q_object
-                self.const = True
-                self.type = "cte"
-            elif op_type == 1:
-                raise Exception("The Qobj must not already be a function")
-            elif op_type == -1:
-                pass
+        if isinstance(Q_object, Qobj):
+            self.cte = Q_object
+            self.const = True
+            self.type = "cte"
         else:
-            op_type_count = [0, 0, 0, 0]
-            for type_, op in zip(op_type, Q_object):
-                if type_ == 0:
+            for op in Q_object:
+                if isinstance(op, Qobj):
                     if self.cte is None:
                         self.cte = op
                     else:
                         self.cte += op
-                elif type_ == 1:
-                    op_type_count[0] += 1
-                    self.ops.append(EvoElement(op[0], op[1], op[1], "func"))
-                elif type_ == 2:
-                    op_type_count[1] += 1
-                    self.ops.append(EvoElement(op[0], _StrWrapper(op[1]),
-                                    op[1], "string"))
-                elif type_ == 3:
-                    op_type_count[2] += 1
-                    self.ops.append(EvoElement(
-                        op[0],
-                        _CubicSplineWrapper(tlist, op[1], args=self.args),
-                        op[1].copy(), "array"))
-                elif type_ == 4:
-                    op_type_count[3] += 1
-                    self.ops.append(EvoElement(op[0], op[1], op[1], "spline"))
+                else:
+                    self.ops.append(EvoElement(op[0], Coeff(op[1])))
 
-            nops = sum(op_type_count)
-            if all([op_t == 0 for op_t in op_type]):
-                self.type = "cte"
-            elif op_type_count[0] == nops:
-                self.type = "func"
-            elif op_type_count[1] == nops:
-                self.type = "string"
-            elif op_type_count[2] == nops:
-                self.type = "array"
-            elif op_type_count[3] == nops:
-                self.type = "spline"
-            elif op_type_count[0]:
-                self.type = "mixed_callable"
-            else:
-                self.type = "mixed_compilable"
+            if self.cte is None:
+                self.cte = self.ops[0].qobj * 0
+                self.dummy_cte = True
 
             try:
-                if not self.cte:
-                    self.cte = self.ops[0].qobj
-                    # test is all qobj are compatible (shape, dims)
-                    for op in self.ops[1:]:
-                        self.cte += op.qobj
-                    self.cte *= 0.
-                    self.dummy_cte = True
-                else:
-                    cte_copy = self.cte.copy()
-                    # test is all qobj are compatible (shape, dims)
-                    for op in self.ops:
-                        cte_copy += op.qobj
+                cte_copy = self.cte.copy()
+                # test is all qobj are compatible (shape, dims)
+                for op in self.ops:
+                    cte_copy += op.qobj
             except Exception as e:
                 raise TypeError("Qobj not compatible.") from e
 
             if not self.ops:
                 self.const = True
-        self.num_obj = (len(self.ops) if self.dummy_cte else len(self.ops) + 1)
         self._args_checks()
         if e_ops:
             for i, dargs in enumerate(self.dynamics_args):
@@ -573,43 +405,6 @@ class QobjEvo:
                                              QobjEvo(e_ops[dargs[2]]))
         if state0 is not None:
             self._dynamics_args_update(0., state0)
-
-    def _td_format_check_single(self, Q_object, tlist=None):
-        op_type = []
-
-        if isinstance(Q_object, Qobj):
-            op_type = 0
-        elif isinstance(Q_object, (FunctionType,
-                                   BuiltinFunctionType, partial)):
-            op_type = 1
-        elif isinstance(Q_object, list):
-            if (len(Q_object) == 0):
-                op_type = -1
-            for op_k in Q_object:
-                if isinstance(op_k, Qobj):
-                    op_type.append(0)
-                elif isinstance(op_k, list):
-                    if not isinstance(op_k[0], Qobj):
-                        raise TypeError("Incorrect Q_object specification")
-                    elif len(op_k) == 2:
-                        if isinstance(op_k[1], Cubic_Spline):
-                            op_type.append(4)
-                        elif callable(op_k[1]):
-                            op_type.append(1)
-                        elif isinstance(op_k[1], str):
-                            op_type.append(2)
-                        elif isinstance(op_k[1], np.ndarray):
-                            if not isinstance(tlist, np.ndarray) or not \
-                                        len(op_k[1]) == len(tlist):
-                                raise TypeError("Time list does not match")
-                            op_type.append(3)
-                        else:
-                            raise TypeError("Incorrect Q_object specification")
-                    else:
-                        raise TypeError("Incorrect Q_object specification")
-        else:
-            raise TypeError("Incorrect Q_object specification")
-        return op_type
 
     def _args_checks(self):
         statedims = [self.cte.dims[1],[1]]
@@ -644,13 +439,7 @@ class QobjEvo:
     def _check_old_with_state(self):
         add_vec = False
         for op in self.ops:
-            if op.type == "func":
-                try:
-                    op.get_coeff(0., self.args)
-                except TypeError as e:
-                    nfunc = _StateAsArgs(self.coeff)
-                    op = EvoElement((op.qobj, nfunc, nfunc, "func"))
-                    add_vec = True
+            add_vec += op.coeff._check_old_with_state()
         if add_vec:
             self.dynamics_args += [("_state_vec", "vec", None)]
 
@@ -680,15 +469,16 @@ class QobjEvo:
         elif data:
             op_t = self.cte.data.copy()
             for part in self.ops:
-                op_t += part.qobj.data * part.get_coeff(t, self.args)
+                op_t += part.qobj.data * part.coeff(t, self.args)
         else:
             op_t = self.cte.copy()
             for part in self.ops:
-                op_t += part.qobj * part.get_coeff(t, self.args)
+                op_t += part.qobj * part.coeff(t, self.args)
 
         return op_t
 
     def _dynamics_args_update(self, t, state):
+        # TO-DO dyn_args class
         if isinstance(state, Qobj):
             for name, what, op in self.dynamics_args:
                 if what == "vec":
@@ -769,12 +559,7 @@ class QobjEvo:
         new.safePickle = self.safePickle
 
         for op in self.ops:
-            if op.type == "array":
-                new_coeff = op.coeff.copy()
-            else:
-                new_coeff = op.coeff
-            new.ops.append(EvoElement(op.qobj.copy(), op.get_coeff,
-                                      new_coeff, op.type))
+            new.ops.append( EvoElement(op.qobj.copy(), op.coeff.copy()) )
 
         return new
 
@@ -798,12 +583,7 @@ class QobjEvo:
         self.safePickle = other.safePickle
 
         for op in other.ops:
-            if op.type == "array":
-                new_coeff = op.coeff.copy()
-            else:
-                new_coeff = op.coeff
-            self.ops.append(EvoElement(op.qobj.copy(), op.get_coeff,
-                                       new_coeff, op.type))
+            self.ops.append( EvoElement(op.qobj.copy(), op.coeff.copy()) )
 
     def arguments(self, new_args):
         if not isinstance(new_args, dict):
@@ -838,6 +618,10 @@ class QobjEvo:
             list_qobj.append([op.qobj, op.coeff])
         return list_qobj
 
+    @property
+    def num_ops(self):
+        return len(self.ops) + (0 if self.dummy_cte else 1)
+
     # Math function
     def __add__(self, other):
         res = self.copy()
@@ -852,67 +636,40 @@ class QobjEvo:
     def __iadd__(self, other):
         if isinstance(other, QobjEvo):
             self.cte += other.cte
-            l = len(self.ops)
             for op in other.ops:
-                if op.type == "array":
-                    new_coeff = op.coeff.copy()
-                else:
-                    new_coeff = op.coeff
-                self.ops.append(EvoElement(op.qobj.copy(), op.get_coeff,
-                                           new_coeff, op.type))
-                l += 1
+                self.ops.append(EvoElement(op.qobj.copy(),
+                                           op.get_coeff.copy() ))
             self.args.update(**other.args)
             self.dynamics_args += other.dynamics_args
             self.const = self.const and other.const
             self.dummy_cte = self.dummy_cte and other.dummy_cte
-            if self.type != other.type:
-                if self.type in ["func", "mixed_callable"] or \
-                        other.type in ["func", "mixed_callable"]:
-                    self.type = "mixed_callable"
-                else:
-                    self.type = "mixed_compilable"
             self.compiled = ""
             self.compiled_qobjevo = None
             self.coeff_get = None
-
-            if self.tlist is None:
-                self.tlist = other.tlist
-            else:
-                if other.tlist is None:
-                    pass
-                elif len(other.tlist) != len(self.tlist) or \
-                        other.tlist[-1] != self.tlist[-1]:
-                    raise Exception("tlist are not compatible")
-            self._reset_type()
         else:
             self.cte += other
             self.dummy_cte = False
             self._recompile()
 
-        self.num_obj = (len(self.ops) if self.dummy_cte else len(self.ops) + 1)
         return self
 
     def __sub__(self, other):
         res = self.copy()
         res -= other
-        self._recompile()
         return res
 
     def __rsub__(self, other):
         res = -self.copy()
         res += other
-        self._recompile()
         return res
 
     def __isub__(self, other):
         self += (-other)
-        self._recompile()
         return self
 
     def __mul__(self, other):
         res = self.copy()
         res *= other
-        self._recompile()
         return res
 
     def __rmul__(self, other):
@@ -950,22 +707,20 @@ class QobjEvo:
                 old_ops = self.ops
                 if not other.dummy_cte:
                     for op in old_ops:
-                        new_terms.append(self._ops_mul_cte(op, other.cte, "R"))
+                        new_terms.append(EvoElement(op.qobj * other.cte,
+                                                    op.coeff))
                 if not self.dummy_cte:
                     for op in other.ops:
-                        new_terms.append(self._ops_mul_cte(op, cte, "L"))
-
-                for op_left in old_ops:
-                    for op_right in other.ops:
-                        new_terms.append(self._ops_mul_(op_left,
-                                                        op_right))
+                        new_terms.append(EvoElement(cte * op.qobj,
+                                                    op.coeff))
+                for op_L in old_ops:
+                    for op_R in other.ops:
+                        new_terms.append(EvoElement(op_L.qobj * op_R.qobj,
+                                                    op_L.coeff * op_R.coeff))
                 self.ops = new_terms
                 self.args.update(other.args)
                 self.dynamics_args += other.dynamics_args
                 self.dummy_cte = self.dummy_cte and other.dummy_cte
-                self.num_obj = (len(self.ops) if
-                              self.dummy_cte else len(self.ops) + 1)
-                self._reset_type()
 
         else:
             raise TypeError("QobjEvo can only be multiplied"
@@ -996,57 +751,31 @@ class QobjEvo:
     def __neg__(self):
         res = self.copy()
         res.cte = -res.cte
-        for op in res.ops:
-            op.qobj = -op.qobj
-        self._recompile()
+        res.ops = [EvoElement(-op.qobj, op.coeff) for op in res.ops]
+        res._recompile()
         return res
-
-    def _ops_mul_(self, opL, opR):
-        new_f = _Prod(opL.get_coeff, opR.get_coeff)
-        new_op = [opL.qobj*opR.qobj, new_f, None, 0]
-        if opL.type == opR.type and opL.type == "string":
-            new_op[2] = "(" + opL.coeff + ") * (" + opR.coeff + ")"
-            new_op[3] = "string"
-        elif opL[3] == opR[3] and opL[3] == "array":
-            new_op[2] = opL[2]*opR[2]
-            new_op[3] = "array"
-        else:
-            new_op[2] = new_f
-            new_op[3] = "func"
-            if self.type not in ["func", "mixed_callable"]:
-                self.type = "mixed_callable"
-        return EvoElement.make(new_op)
-
-    def _ops_mul_cte(self, op, cte, side):
-        new_op = [None, op.get_coeff, op.coeff, op.type]
-        if side == "R":
-            new_op[0] = op.qobj * cte
-        if side == "L":
-            new_op[0] = cte * op.qobj
-        return EvoElement.make(new_op)
 
     # Transformations
     def trans(self):
         res = self.copy()
         res.cte = res.cte.trans()
-        for op in res.ops:
-            op.qobj = op.qobj.trans()
+        res.ops = [EvoElement(op.qobj.trans(), op.coeff)
+                   for op in self.ops)]
+        res._recompile()
         return res
 
     def conj(self):
         res = self.copy()
         res.cte = res.cte.conj()
-        for op in res.ops:
-            op.qobj = op.qobj.conj()
-        res._f_conj()
+        res.ops = [EvoElement(op.qobj.conj(), op.coeff.conj())
+                   for op in self.ops)]
         return res
 
     def dag(self):
         res = self.copy()
         res.cte = res.cte.dag()
-        for op in res.ops:
-            op.qobj = op.qobj.dag()
-        res._f_conj()
+        res.ops = [EvoElement(op.qobj.dag(), op.coeff.conj())
+                   for op in self.ops)]
         return res
 
     def _cdc(self):
@@ -1057,10 +786,16 @@ class QobjEvo:
         else:
             res = self.copy()
             res.cte = res.cte.dag() * res.cte
-            for op in res.ops:
-                op.qobj = op.qobj.dag() * op.qobj
-            res._f_norm2()
+            res.ops = [EvoElement(op.qobj.dag() * op.qobj, op.coeff._cdc())
+                       for op in self.ops)]
         return res
+
+    def _shift(self):
+        self.compiled = ""
+        self.coeff_get = None
+        self.args.update({"_t0": 0})
+        self.ops = [EvoElement(op.qobj, op.coeff._shift()) for op in self.ops)]
+        return self
 
     # Unitary function of Qobj
     def tidyup(self, atol=1e-12):
@@ -1073,36 +808,22 @@ class QobjEvo:
         sets = []
         callable_flags = ["func", "spline"]
         for i, op1 in enumerate(self.ops):
-            already_matched = False
-            for _set in sets:
-                already_matched = already_matched or i in _set
+            already_matched = any(i in _set for _set in sets)
             if not already_matched:
                 this_set = [i]
                 for j, op2 in enumerate(self.ops[i+1:]):
                     if op1.qobj == op2.qobj:
-                        same_flag = op1.type == op2.type
-                        callable_1 = op1.type in callable_flags
-                        callable_2 = op2.type in callable_flags
-                        if (same_flag or (callable_1 and callable_2)):
-                            this_set.append(j+i+1)
+                        this_set.append(j+i+1)
                 sets.append(this_set)
 
         fsets = []
         for i, op1 in enumerate(self.ops):
-            already_matched = False
-            for _set in fsets:
-                already_matched = already_matched or i in _set
+            already_matched = any(i in _set for _set in fsets)
             if not already_matched:
                 this_set = [i]
                 for j, op2 in enumerate(self.ops[i+1:]):
-                    if op1.type != op2.type:
-                        pass
-                    elif op1.type is "array":
-                        if np.allclose(op1.coeff, op2.coeff):
-                            this_set.append(j+i+1)
-                    else:
-                        if op1.coeff is op2.coeff:
-                            this_set.append(j+i+1)
+                    if op1.coeff is op2.coeff:
+                        this_set.append(j+i+1)
                 fsets.append(this_set)
         return sets, fsets
 
@@ -1113,33 +834,13 @@ class QobjEvo:
             if len(_set) == 1:
                 new_ops.append(self.ops[_set[0]])
 
-            elif self.ops[_set[0]].type in callable_flags:
-                new_op = [self.ops[_set[0]].qobj, None, None, "func"]
-                fs = []
+            else:
+                # Would be nice but need to defined addition with 0 (int)
+                # new_coeff = sum(self.ops[i].coeff for i in _set)
+                new_coeff = self.ops[_set[0]].coeff
                 for i in _set:
-                    fs += [self.ops[i].get_coeff]
-                new_op[1] = _Add(fs)
-                new_op[2] = new_op[1]
-                new_ops.append(EvoElement.make(new_op))
-
-            elif self.ops[_set[0]].type is "string":
-                new_op = [self.ops[_set[0]].qobj, None, None, "string"]
-                new_str = "(" + self.ops[_set[0]].coeff + ")"
-                for i in _set[1:]:
-                    new_str += " + (" + self.ops[i].coeff + ")"
-                new_op[1] = _StrWrapper(new_str)
-                new_op[2] = new_str
-                new_ops.append(EvoElement.make(new_op))
-
-            elif self.ops[_set[0]].type is "array":
-                new_op = [self.ops[_set[0]].qobj, None, None, "array"]
-                new_array = (self.ops[_set[0]].coeff).copy()
-                for i in _set[1:]:
-                    new_array += self.ops[i].coeff
-                new_op[2] = new_array
-                new_op[1] = _CubicSplineWrapper(
-                    self.tlist, new_array, args=self.args)
-                new_ops.append(EvoElement.make(new_op))
+                    new_coeff += self.ops[i].coeff
+                new_ops.append(EvoElement(self.ops[_set[0]].qobj, new_coeff))
 
         self.ops = new_ops
 
@@ -1147,14 +848,16 @@ class QobjEvo:
         new_ops = []
         for _set in fsets:
             base = self.ops[_set[0]]
-            new_op = [None, base.get_coeff, base.coeff, base.type]
+            new_op = [None, base.coeff]
+            new_op[0] += sum(self.ops[i].qobj for
+
             if len(_set) == 1:
                 new_op[0] = base.qobj
             else:
                 new_op[0] = base.qobj.copy()
                 for i in _set[1:]:
                     new_op[0] += self.ops[i].qobj
-            new_ops.append(EvoElement.make(new_op))
+            new_ops.append(EvoElement(*new_op))
         self.ops = new_ops
 
     def compress(self):
@@ -1183,39 +886,6 @@ class QobjEvo:
             self._compress_merge_func(fsets)
             self._reset_type()
 
-    def _reset_type(self):
-        self.compiled = ""
-        self.compiled_qobjevo = None
-        self.coeff_get = None
-        op_type_count = [0, 0, 0, 0]
-        for op in self.ops:
-            if op.type == "func":
-                op_type_count[0] += 1
-            elif op.type == "string":
-                op_type_count[1] += 1
-            elif op.type == "array":
-                op_type_count[2] += 1
-            elif op.type == "spline":
-                op_type_count[3] += 1
-
-        nops = sum(op_type_count)
-        if not self.ops and self.dummy_cte is False:
-            self.type = "cte"
-        elif op_type_count[0] == nops:
-            self.type = "func"
-        elif op_type_count[1] == nops:
-            self.type = "string"
-        elif op_type_count[2] == nops:
-            self.type = "array"
-        elif op_type_count[3] == nops:
-            self.type = "spline"
-        elif op_type_count[0]:
-            self.type = "mixed_callable"
-        else:
-            self.type = "mixed_compilable"
-
-        self.num_obj = (len(self.ops) if self.dummy_cte else len(self.ops) + 1)
-
     def permute(self, order):
         res = self.copy()
         res.cte = res.cte.permute(order)
@@ -1235,132 +905,6 @@ class QobjEvo:
         if self.compiled:
             self._recompile()
         return res
-
-    def apply_decorator(self, function, *args, **kw_args):
-        if "str_mod" in kw_args:
-            str_mod = kw_args["str_mod"]
-            del kw_args["str_mod"]
-        else:
-            str_mod = None
-        if "inplace_np" in kw_args:
-            inplace_np = kw_args["inplace_np"]
-            del kw_args["inplace_np"]
-        else:
-            inplace_np = None
-        res = self.copy()
-        for op in res.ops:
-            op.get_coeff = function(op.get_coeff, *args, **kw_args)
-            if op.type == ["func", "spline"]:
-                op.coeff = op.get_coeff
-                op.type = "func"
-            elif op.type == "string":
-                if str_mod is None:
-                    op.coeff = op.get_coeff
-                    op.type = "func"
-                else:
-                    op.coeff = str_mod[0] + op.coeff + str_mod[1]
-            elif op.type == "array":
-                if inplace_np:
-                    # keep the original function, change the array
-                    def f(a):
-                        return a
-                    ff = function(f, *args, **kw_args)
-                    for i, v in enumerate(op.coeff):
-                        op.coeff[i] = ff(v)
-                    op.get_coeff = _CubicSplineWrapper(
-                        self.tlist, op.coeff, args=self.args)
-                else:
-                    op.coeff = op.get_coeff
-                    op.type = "func"
-        if self.type == "string" and str_mod is None:
-            res.type = "mixed_callable"
-        elif self.type == "array" and not inplace_np:
-            res.type = "mixed_callable"
-        elif self.type == "spline":
-            res.type = "mixed_callable"
-        elif self.type == "mixed_compilable":
-            for op in res.ops:
-                if op.type == "func":
-                    res.type = "mixed_callable"
-        return res
-
-    def _f_norm2(self):
-        self.compiled = ""
-        self.coeff_get = None
-        new_ops = []
-        for op in self.ops:
-            new_op = [op.qobj, None, None, op.type]
-            if op.type == "func":
-                new_op[1] = _Norm2(op.get_coeff)
-                new_op[2] = new_op[1]
-            elif op.type == "string":
-                new_op[2] = "norm(" + op.coeff + ")"
-                new_op[1] = _StrWrapper(new_op[2])
-            elif op.type == "array":
-                new_op[2] = np.abs(op.coeff)**2
-                new_op[1] = _CubicSplineWrapper(
-                    self.tlist, new_op[2], args=self.args)
-            elif op.type == "spline":
-                new_op[1] = _Norm2(op.get_coeff)
-                new_op[2] = new_op[1]
-                new_op[3] = "func"
-                self.type = "mixed_callable"
-            new_ops.append(EvoElement.make(new_op))
-        self.ops = new_ops
-        return self
-
-    def _f_conj(self):
-        self.compiled = ""
-        self.coeff_get = None
-        new_ops = []
-        for op in self.ops:
-            new_op = [op.qobj, None, None, op.type]
-            if op.type == "func":
-                new_op[1] = _Conj(op.get_coeff)
-                new_op[2] = new_op[1]
-            elif op.type == "string":
-                new_op[2] = "conj(" + op.coeff + ")"
-                new_op[1] = _StrWrapper(new_op[2])
-            elif op.type == "array":
-                new_op[2] = np.conj(op.coeff)
-                new_op[1] = _CubicSplineWrapper(
-                    self.tlist, new_op[2], args=self.args)
-            elif op.type == "spline":
-                new_op[1] = _Conj(op.get_coeff)
-                new_op[2] = new_op[1]
-                new_op[3] = "func"
-                self.type = "mixed_callable"
-            new_ops.append(EvoElement.make(new_op))
-        self.ops = new_ops
-        return self
-
-    def _shift(self):
-        self.compiled = ""
-        self.coeff_get = None
-        self.args.update({"_t0": 0})
-        new_ops = []
-        for op in self.ops:
-            new_op = [op.qobj, None, None, op.type]
-            if op.type == "func":
-                new_op[1] = _Shift(op.get_coeff)
-                new_op[2] = new_op[1]
-            elif op.type == "string":
-                new_op[2] = sub("(?<=[^0-9a-zA-Z_])t(?=[^0-9a-zA-Z_])",
-                                "(t+_t0)", " " + op.coeff + " ")
-                new_op[1] = _StrWrapper(new_op[2])
-            elif op.type == "array":
-                new_op[2] = _Shift(op.get_coeff)
-                new_op[1] = new_op[1]
-                new_op[3] = "func"
-                self.type = "mixed_callable"
-            elif op.type == "spline":
-                new_op[1] = _Shift(op.get_coeff)
-                new_op[2] = new_op[1]
-                new_op[3] = "func"
-                self.type = "mixed_callable"
-            new_ops.append(EvoElement.make(new_op))
-        self.ops = new_ops
-        return self
 
     def expect(self, t, state, herm=0):
         if not isinstance(t, (int, float)):
@@ -1494,91 +1038,16 @@ class QobjEvo:
         if code:
             return Code
 
+    def _compile_coeff(self):
+        coeffs = [op.coeff for op in self.ops]
+        self._get_coeff, code = compile_united_coeff(coeffs)
+        return code
+
     def _recompile(self):
         if self.compiled:
             self.compiled_qobjevo.set_data(self.cte, self.ops)
             self.compiled_qobjevo.has_dyn_args(bool(self.dynamics_args))
             self.compiled_qobjevo.set_num_threads(self.omp)
-
-    def _compile_coeff(self):
-        Code = None
-        if self.type in ["func"]:
-            funclist = [part.get_coeff for part in self.ops]
-            self.coeff_get = _UnitedFuncCaller(funclist, self.args,
-                                               self.dynamics_args,
-                                               self.cte)
-        elif self.type in ["mixed_callable"] and self.use_cython:
-            funclist = []
-            for part in self.ops:
-                if isinstance(part.get_coeff, _StrWrapper):
-                    get_coeff, file_ = _compile_str_single(part.coeff,
-                                                           self.args)
-                    coeff_files.add(file_)
-                    funclist.append(get_coeff)
-                else:
-                    funclist.append(part.get_coeff)
-
-            self.coeff_get = _UnitedFuncCaller(funclist, self.args,
-                                               self.dynamics_args,
-                                               self.cte)
-        elif self.type in ["mixed_callable"]:
-            funclist = [part.get_coeff for part in self.ops]
-            _UnitedStrCaller, Code, file_ = _compiled_coeffs_python(
-                                                    self.ops,
-                                                    self.args,
-                                                    self.dynamics_args,
-                                                    self.tlist)
-            coeff_files.add(file_)
-            self.coeff_get = _UnitedStrCaller(funclist, self.args,
-                                              self.dynamics_args,
-                                              self.cte)
-        elif self.type in ["string", "mixed_compilable"]:
-            if self.use_cython:
-                # All factor can be compiled
-                self.coeff_get, Code, file_ = _compiled_coeffs(
-                                                    self.ops,
-                                                    self.args,
-                                                    self.dynamics_args,
-                                                    self.tlist)
-                coeff_files.add(file_)
-            else:
-                # All factor can be compiled
-                _UnitedStrCaller, Code, file_ = _compiled_coeffs_python(
-                                                    self.ops,
-                                                    self.args,
-                                                    self.dynamics_args,
-                                                    self.tlist)
-                coeff_files.add(file_)
-                funclist = [part.get_coeff for part in self.ops]
-                self.coeff_get = _UnitedStrCaller(funclist, self.args,
-                                                  self.dynamics_args,
-                                                  self.cte)
-
-        elif self.type == "array":
-            try:
-                use_step_func = self.args["_step_func_coeff"]
-            except KeyError:
-                use_step_func = 0
-            if np.allclose(np.diff(self.tlist),
-                    self.tlist[1] - self.tlist[0]):
-                if use_step_func:
-                    self.coeff_get = StepCoeffCte(
-                        self.ops, None, self.tlist)
-                else:
-                    self.coeff_get = InterCoeffCte(
-                        self.ops, None, self.tlist)
-            else:
-                if use_step_func:
-                    self.coeff_get = StepCoeffT(
-                        self.ops, None, self.tlist)
-                else:
-                    self.coeff_get = InterCoeffT(
-                        self.ops, None, self.tlist)
-
-        elif self.type == "spline":
-            self.coeff_get = InterpolateCoeff(self.ops, None, None)
-
-        return Code
 
     def _get_coeff(self, t):
         out = []
@@ -1619,105 +1088,6 @@ class QobjEvo:
             self.compiled_qobjevo = \
                 CQobjEvoTdMatched.__new__(CQobjEvoTdMatched)
             self.compiled_qobjevo.__setstate__(state[1])
-
-
-# Function defined inside another function cannot be pickled,
-# Using class instead
-class _UnitedFuncCaller:
-    def __init__(self, funclist, args, dynamics_args, cte):
-        self.funclist = funclist
-        self.args = args
-        self.dynamics_args = dynamics_args
-        self.dims = cte.dims
-        self.shape = cte.shape
-
-    def set_args(self, args, dynamics_args=[]):
-        self.args = args
-        if dynamics_args:
-            self.dynamics_args = dynamics_args
-
-    def dyn_args(self, t, state, shape):
-        # 1d array are to F ordered
-        mat = state.reshape(shape, order="F")
-        for name, what, op in self.dynamics_args:
-            if what == "vec":
-                self.args[name] = state
-            elif what == "mat":
-                self.args[name] = mat
-            elif what == "Qobj":
-                if self.shape[1] == shape[1]:  # oper
-                    self.args[name] = Qobj(mat, dims=self.dims)
-                elif shape[1] == 1: # ket
-                    self.args[name] = Qobj(mat, dims=[self.dims[1],[1]])
-                else:  # rho
-                    self.args[name] = Qobj(mat, dims=self.dims[1])
-            elif what == "expect":
-                if shape[1] == op.cte.shape[1]: # same shape as object
-                    self.args[name] = op.mul_mat(t, mat).trace()
-                else:
-                    self.args[name] = op.expect(t, state)
-
-    def __call__(self, t, args={}):
-        if args:
-            now_args = self.args.copy()
-            now_args.update(args)
-        else:
-            now_args = self.args
-        out = []
-        for func in self.funclist:
-            out.append(func(t, now_args))
-        return out
-
-    def get_args(self):
-        return self.args
-
-    def copy(self):
-        res = _UnitedFuncCaller(self.funclist,
-                                self.args,
-                                self.dynamics_args,
-                                self)
-
-
-class _Norm2():
-    def __init__(self, f):
-        self.func = f
-
-    def __call__(self, t, args):
-        return self.func(t, args)*np.conj(self.func(t, args))
-
-
-class _Shift():
-    def __init__(self, f):
-        self.func = f
-
-    def __call__(self, t, args):
-        return np.conj(self.func(t + args["_t0"], args))
-
-
-class _Conj():
-    def __init__(self, f):
-        self.func = f
-
-    def __call__(self, t, args):
-        return np.conj(self.func(t, args))
-
-
-class _Prod():
-    def __init__(self, f, g):
-        self.func_1 = f
-        self.func_2 = g
-
-    def __call__(self, t, args):
-        return self.func_1(t, args)*self.func_2(t, args)
-
-
-class _Add():
-    def __init__(self, fs):
-        self.funcs = fs
-
-    def __call__(self, t, args):
-        return np.sum([f(t, args) for f in self.funcs])
-
 
 from qutip.superoperator import vec2mat
 from qutip.cy.cqobjevo import (CQobjEvoTd, CQobjEvoTdMatched, CQobjEvoTdDense)
